requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
inflection==0.5.1
jsonschema==4.17.3
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
